# swagger/openapi.yaml
...
openapi: 3.0.3
info:
  title: Task Manager API
  description: |
    Production-grade REST API with JWT Authentication, Role-Based Access Control,
    and full CRUD operations for Task management.

    ## Authentication
    - Register or login to receive an **access token**
    - Pass the token in the `Authorization` header as `Bearer <token>`
    - Access tokens expire in **15 minutes**
    - Use `POST /auth/refresh` to get a new access token via httpOnly cookie

    ## Roles
    - **USER** — Can manage only their own tasks
    - **ADMIN** — Can manage all tasks across all users
  version: 1.0.0
  contact:
    name: Gurrala Sai Haneesh
    email: joydip@primetrade.ai
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful.
        data:
          type: object

    # ── Error Response Envelope ────────────────
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred.
        errorCode:
          type: string
          example: VALIDATION_ERROR
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── User ───────────────────────────────────
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Issues a new access token using the httpOnly refresh token cookie.
      responses:
        "200":
          description: Access token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
        "401":
          description: No refresh token or token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...